intern_xplorer_api/requirements.txt
psycopg2
django
djangorestframework
django-extensions
django-elasticsearch-dsl-drf
